updated code for vpc-routetables
