+++ aws-classic-ts-vpc-routetables-no-loops/Pulumi.dev.yaml
@@ -1,0 +1,2 @@
+config:
+  aws:region: us-east-2
