+++ aws-classic-ts-vpc-routetables-no-loops/Pulumi.yaml
@@ -1,0 +1,3 @@
+name: aws-classic-ts-vpc-routetables-no-loops
+description: A minimal AWS TypeScript Pulumi program
+runtime: nodejs
