+++ aws-classic-ts-vpc-routetables-no-loops/package.json
@@ -1,0 +1,11 @@
+{
+    "name": "aws-classic-ts-vpc-routetables-no-loops",
+    "main": "index.ts",
+    "devDependencies": {
+        "@types/node": "^14"
+    },
+    "dependencies": {
+        "@pulumi/pulumi": "^3.0.0",
+        "@pulumi/aws": "^5.0.0"
+    }
+}
